fixes varios

+++ tpEspecial/mantenimiento-service/src/main/java/edu/tudai/arq/mantenimientoservice/feignclient/MonopatinFeignClient.java
@@ -26,8 +26,8 @@ public interface MonopatinFeignClient {
     record MonopatinResponse(
             Long id,
             String estado,
-            Double latitud,
-            Double longitud,
+            Integer latitud,
+            Integer longitud,
             Double kilometrosTotales,
             Long tiempoUsoTotal
     ) {}
